Restructured the test folder
Updated make file commands.

Removed unused imports

Added singleton test to unit tests.
