+++ tests/property/test_hash.py
@@ -4,7 +4,7 @@ from hypothesis import given
 
 from electionguard.group import ElementModQ
 from electionguard.hash import hash_elems
-from tests.test_group import elements_mod_p, elements_mod_q
+from tests.property.test_group import elements_mod_p, elements_mod_q
 
 
 class TestHash(unittest.TestCase):
