+++ tests/property/test_tally.py
