+++ tests/unit/test_election_polynomial.py
