+++ tests/unit/test_singleton.py
